extract rate limiting logic into a reusable utility
